chore: hermetic tests, shared value helper and coverage gates
Fixes the backend test suite reading the shared user cache directory.
`ignore_datasets_file` defaulted to a platformdirs path written by any
local REF run, so a stale file there failed all 51 API tests locally
while CI stayed green. Tests now point at a file under the test data.

- Extracts the duplicated metric-value endpoint body into
  `fetch_metric_values`. The diagnostic-scoped and execution-scoped
  routes only differed in the filter and the CSV filename.
- Keys the diagnostic metadata cache by path via an `lru_cache` instead
  of holding it on `DiagnosticSummary` as a class variable.
- Splits `ref_backend.models` into a package. The generated OpenAPI
  schema is unchanged.
- Removes `SECRET_KEY`, which was only ever read to warn that it was
  still set to the default.
- Adds v8 coverage reporting to the frontend, wired into CI with
  thresholds set at the current numbers.
- Raises the backend coverage gate from 50 to 90 percent, which the
  suite already meets now that it passes.
