+++ .env.sample
@@ -12,7 +12,6 @@ PROJECT_NAME='Rapid Evaluation Framework'
 BACKEND_CORS_ORIGINS="http://localhost,http://localhost:5173,https://localhost,https://localhost:5173"
 # Regex pattern for dynamic CORS origins (e.g. Netlify deploy previews)
 # BACKEND_CORS_ORIGIN_REGEX="https://deploy-preview-\d+--climate-ref\.netlify\.app"
-SECRET_KEY=changethis
 
 # Error tracking
 SENTRY_DSN=
