+++ .github/workflows/ci.yml
@@ -54,8 +54,8 @@ jobs:
     - name: Run Biome
       run: biome ci .
 
-    - name: Run tests
-      run: npm run test
+    - name: Run tests with coverage
+      run: npm run test:coverage
 
     - name: Build
       run: npm run build
@@ -91,4 +91,4 @@ jobs:
 
     - name: Run tests
       run: |
-        uv run coverage run -m pytest tests/ && uv run coverage report --fail-under=50
+        uv run coverage run -m pytest tests/ && uv run coverage report --fail-under=90
