+++ .gitignore
@@ -3,3 +3,5 @@
 
 .pytest_cache
 .ruff_cache
+
+frontend/coverage/
