+++ frontend/package.json
@@ -9,7 +9,8 @@
     "lint": "biome check --write .",
     "preview": "vite preview",
     "openapi-ts": "openapi-ts",
-    "test": "vitest run"
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
     "@floating-ui/react": "^0.27.16",
@@ -69,6 +70,7 @@
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
     "@vitejs/plugin-react-swc": "^3.8.0",
+    "@vitest/coverage-v8": "^4.1.2",
     "@welldone-software/why-did-you-render": "^10.0.1",
     "globals": "^15.15.0",
     "jsdom": "^28.1.0",
